requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.16
drf-yasg>=1.21
mysqlclient>=2.2


